added support for summarizing youtube videos

+++ main.py
@@ -4,31 +4,51 @@ import streamlit as st
 from utils import (
     doc_loader, summary_prompt_creator, doc_to_final_summary
 )
-from my_prompts import map_prompt, combine_prompt
+from my_prompts import file_map, file_combine, youtube_map, youtube_combine
 from streamlit_app_utils import check_gpt_4, check_key_validity, create_temp_file, create_chat_model, \
     token_limit, token_minimum
 
+from youtube import transcript_loader
+
 
 def main():
+    """
+    The main function for the Streamlit app.
+
+    :return: None.
+    """
     st.title("Document Summarizer")
-    uploaded_file = st.file_uploader("Upload a document to summarize, 10k to 100k tokens works best!", type=['txt', 'pdf'])
+
+    input_method = st.radio("Select input method", ('Upload a document', 'Enter a YouTube URL'))
+
+    if input_method == 'Upload a document':
+        uploaded_file = st.file_uploader("Upload a document to summarize, 10k to 100k tokens works best!", type=['txt', 'pdf'])
+
+    if input_method == 'Enter a YouTube URL':
+        youtube_url = st.text_input("Enter a YouTube URL to summarize")
+
     api_key = st.text_input("Free usage cap has been hit! Enter your own API key here, or contact the author if you don't have one.")
     st.markdown('[Author email](mailto:ethanujohnston@gmail.com)')
     use_gpt_4 = st.checkbox("Use GPT-4 for the final prompt (STRONGLY recommended, requires GPT-4 API access)", value=True)
-    find_clusters = st.checkbox('Find optimal clusters (experimental, may take a while, but saves on tokens)', value=False)
+    find_clusters = st.checkbox('Find optimal clusters (experimental, could save on token usage)', value=False)
     st.sidebar.markdown('# Made by: [Ethan](https://github.com/e-johnstonn)')
     st.sidebar.markdown('# Git link: [Docsummarizer](https://github.com/e-johnstonn/docsummarizer)')
 
 
     if st.button('Summarize (click once and wait)'):
-        process_summarize_button(uploaded_file, api_key, use_gpt_4, find_clusters)
+        if input_method == 'Upload a document':
+            process_summarize_button(uploaded_file, api_key, use_gpt_4, find_clusters)
+
+        else:
+            doc = transcript_loader(youtube_url)
+            process_summarize_button(doc, api_key, use_gpt_4, find_clusters, file=False)
 
 
-def process_summarize_button(uploaded_file, api_key, use_gpt_4, find_clusters):
+def process_summarize_button(file_or_transcript, api_key, use_gpt_4, find_clusters, file=True):
     """
     Processes the summarize button, and displays the summary if input and doc size are valid
 
-    :param uploaded_file: The file uploaded by the user
+    :param file_or_transcript: The file uploaded by the user or the transcript from the YouTube URL
 
     :param api_key: The API key entered by the user
 
@@ -38,18 +58,26 @@ def process_summarize_button(uploaded_file, api_key, use_gpt_4, find_clusters):
 
     :return: None
     """
-    if not validate_input(uploaded_file, api_key, use_gpt_4):
+    if not validate_input(file_or_transcript, api_key, use_gpt_4):
         return
 
     with st.spinner("Summarizing... please wait..."):
-        temp_file_path = create_temp_file(uploaded_file)
+        if file:
+            temp_file_path = create_temp_file(file_or_transcript)
+            doc = doc_loader(temp_file_path)
+            map_prompt = file_map
+            combine_prompt = file_combine
+        else:
+            doc = file_or_transcript
+            map_prompt = youtube_map
+            combine_prompt = youtube_combine
         llm = create_chat_model(api_key, use_gpt_4)
         initial_prompt_list = summary_prompt_creator(map_prompt, 'text', llm)
         final_prompt_list = summary_prompt_creator(combine_prompt, 'text', llm)
-        doc = doc_loader(temp_file_path)
 
         if not validate_doc_size(doc):
-            os.unlink(temp_file_path)
+            if file:
+                os.unlink(temp_file_path)
             return
 
         if find_clusters:
@@ -59,7 +87,8 @@ def process_summarize_button(uploaded_file, api_key, use_gpt_4, find_clusters):
             summary = doc_to_final_summary(doc, 10, initial_prompt_list, final_prompt_list, api_key, use_gpt_4)
 
         st.markdown(summary, unsafe_allow_html=True)
-        os.unlink(temp_file_path)
+        if file:
+            os.unlink(temp_file_path)
 
 
 def validate_doc_size(doc):
@@ -71,20 +100,20 @@ def validate_doc_size(doc):
     :return: True if the doc is valid, False otherwise
     """
     if not token_limit(doc, 400000):
-        st.warning('File too big!')
+        st.warning('File or transcript too big!')
         return False
 
     if not token_minimum(doc, 2000):
-        st.warning('File too small!')
+        st.warning('File or transcript too small!')
         return False
     return True
 
 
-def validate_input(uploaded_file, api_key, use_gpt_4):
+def validate_input(file_or_transcript, api_key, use_gpt_4):
     """
     Validates the user input, and displays warnings if the input is invalid
 
-    :param uploaded_file: The file uploaded by the user
+    :param file_or_transcript: The file uploaded by the user or the YouTube URL entered by the user
 
     :param api_key: The API key entered by the user
 
@@ -92,8 +121,8 @@ def validate_input(uploaded_file, api_key, use_gpt_4):
 
     :return: True if the input is valid, False otherwise
     """
-    if uploaded_file is None:
-        st.warning("Please upload a file.")
+    if file_or_transcript == None:
+        st.warning("Please upload a file or enter a YouTube URL.")
         return False
 
     if not check_key_validity(api_key):
